refactoring repositories and services

+++ services/albumItemService.js
@@ -1,4 +1,6 @@
 const Repository = require('../repositories/albumItemRepository');
+const ApiError = require("../utils/ApiError");
+const httpStatus = require("../utils/statusCodes");
 
 class AlbumItemService {
     async createAlbumItem(post_id, album_id) {
@@ -8,6 +10,8 @@ class AlbumItemService {
         return Repository.getAll();
     };
     async deleteAlbumItem(id) {
+        const albumItem = await Repository.getById(id);
+        if (!albumItem) throw new ApiError(httpStatus.NOT_FOUND, 'Album item not found');
         return Repository.delete(id);
     };
 }
